pipeline/requirements.txt
google-play-scraper
anthropic
openai
supabase
python-dotenv
